+++ lib/src/scope.dart
@@ -1,8 +1,8 @@
 part of "../componentt.dart";
 
-/// Provide access to [ComponentMixin] using context
+/// Provide access to [ComponentStateMixin] using context
 class _ComponentScope extends InheritedWidget {
-  final ComponentMixin component;
+  final ComponentStateMixin component;
 
   const _ComponentScope({
     Key? key,
@@ -18,15 +18,15 @@ class _ComponentScope extends InheritedWidget {
     return result!;
   }
 
-  /// Provide successor of [ComponentMixin] from [_ComponentScope] and downcast it to type [C]
+  /// Provide successor of [ComponentStateMixin] from [_ComponentScope] and downcast it to type [C]
   ///
   /// Throws [AssertionError] in debug if [component] is not implement [C]
-  static C componentOf<C extends ComponentMixin>(BuildContext context) {
+  static C componentOf<C extends ComponentStateMixin>(BuildContext context) {
     final component = _of(context).component;
     assert(
       component is C,
-      'Component $C not present in current context. '
-      'Component ${component.runtimeType} available in current context',
+      'ComponentState $C not present in current context. '
+      'ComponentState ${component.runtimeType} available in current context',
     );
     return component as C;
   }