+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.1.0
 
 * [FEAT] Implement ComponentAction API
-* [FEAT] Implement ComponentMixin API necessary to use ComponentActions
+* [FEAT] Implement ComponentStateMixin API necessary to use ComponentActions
 * [FEAT] Implement extensions for BuildContext, ComponentAction and Component 
   handler function
 
@@ -16,5 +16,13 @@
 
 ## 0.3.0
 
-* [FEAT] implement API to access ComponentMixin through BuildContext
+* [FEAT] implement API to access ComponentStateMixin through BuildContext
 * [CHORE] update upper bound of Dart SDK version in pubspec.yaml
+
+## 0.4.0
+
+* [FEAT] Generalized api for creating actions from a handler without context and with context
+* [FEAT] Implement class for alternative component declaration
+* [REFACTORING] Rename ComponentMixin to ComponentStateMixin
+* [REFACTORING] Rename Component to ComponentState
+* [TEST] Improve test coverage